+++ CHANGELOG.md
@@ -9,6 +9,8 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 
 * [mse121682: Links to election nominations don't work after nominations close](http://meta.stackexchange.com/q/121682)
 * [mse230536: Large down-vote count doesn't display negative sign](http://meta.stackexchange.com/q/230536)
+* [math12902: Visited questions are practically indistinguishable in search results](http://meta.math.stackexchange.com/q/12902) (CSS, math.SE only)
+* [math16559: Typo in site CSS disables visited link color in community bulletin](http://meta.math.stackexchange.com/q/16559) (CSS, math.SE only)
 
 1.26 (10 Dec 2014)
 ====